requirements.txt
azure-functions
pyld
pyyaml 
pytest 
requests
aiohttp
